+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - input: 高速なAPIを追加
+- generic: ReadSectionProviderをpubにする
 - eframe: フォント読み込みで大文字小文字違いを許容するようにする
 - eframe: Yu Gothic UIを常に読み込むようにする